requirements.txt
streamlit
pandas
matplotlib
plotly
streamlit-lottie
Pillow
numpy
scipy
